+++ pubspec.yaml
@@ -64,7 +64,9 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - lib/assets/images/background.png
-    - lib/assets/images/callButton.png
+    - lib/assets/images/call_token.png
+    - lib/assets/images/slot_machine.png
+    - lib/assets/images/mini_playing_card.png
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/to/resolution-aware-images
@@ -80,10 +82,10 @@ flutter:
   fonts:
     - family: IntroRust 
       fonts:
-        - asset: lib/assets/fonts/IntroRustG.otf
+        - asset: lib/assets/fonts/rust.otf
     - family: LemonMilkMedium 
       fonts: 
-        - asset: lib/assets/fonts/LEMONMILK-Medium.otf
+        - asset: lib/assets/fonts/lemonMilk.otf
   #
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/to/font-from-package
